+++ src/controller/command/motion/motion.cc
@@ -4,7 +4,7 @@
 
 namespace VM {
     inline Cursor Motion::startPosition(const PtrCursor &cursor) {return cursor;}
-    Motion::Motion(int quantifier) : quantifier (quantifier) {}
+    Motion::Motion(size_t quantifier) : quantifier (quantifier) {}
     Motion::Motion() : quantifier(1) {}
 
     Motion::~Motion() {}
